requirements.txt
chromadb
ollama
python-dotenv
